conversational-knowledge-bot/requirements.txt
streamlit
fastapi
uvicorn
langchain
langchain-groq
python-dotenv
requests
langchain-community
ddgs

